Don't do the Makefile variable checking in "manaliases"
The checking to make sure that the man page being
made is in $(MANALIASES) is something that only
developers need, and in order to be conservative
in installing (e.g., do we have "expr" on Windows?
I have no clue.) move the check to a separate target.
Future work is to add the new "checkaliases" target
to "all" if you're a developer.