NEWS: snmpd: view application filters out unnecessary handlers
There was a bug in the way views were applied to decide whether
or not to skip a given handler - the last segment of the oid would
be stripped, e.g., when deciding whether to traverse into
ifTable (interfaces.2), the view would be checked against
interfaces.  This would result in the ifTable being traversed
(and all results ignored) even if the view only includes
ifNumber and even if it explicitly excludes ifTable.